silence warning from -Wconversion
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/branches/am@2932 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/utils/ncmpidump/vardata.c
@@ -298,14 +298,14 @@ annotate(
       case LANG_C:
 	/* C variable indices */
 	for (id = 0; id < vrank-1; id++)
-	  Printf("%lu,", (unsigned long) cor[id]);
-	Printf("%lu", (unsigned long) cor[id] + iel);
+	  Printf("%lld,", cor[id]);
+	Printf("%lld", cor[id] + iel);
 	break;
       case LANG_F:
 	/* Fortran variable indices */
-	Printf("%lu", (unsigned long) cor[vrank-1] + iel + 1);
+	Printf("%lld", cor[vrank-1] + iel + 1);
 	for (id = vrank-2; id >=0 ; id--) {
-	    Printf(",%lu", 1 + (unsigned long) cor[id]);
+	    Printf(",%lld", 1 + cor[id]);
 	}
 	break;
     }
